+++ muitv.h
@@ -8,7 +8,7 @@ extern "C" __declspec(dllimport) void* muitv_calloc(size_t count, size_t size);
 extern "C" __declspec(dllimport) void* muitv_realloc(void* ptr, size_t size);
 extern "C" __declspec(dllimport) void muitv_free(void* ptr);
 extern "C" __declspec(dllimport) size_t muitv_get_size(void* ptr);
-extern "C" __declspec(dllimport) void muitv_add_call_stack_to_tree(unsigned size);
+extern "C" __declspec(dllimport) void muitv_add_call_stack_to_tree(size_t size);
 
 #if !defined(MUITV_MANUAL_ONLY)
 void* operator new(size_t size)
